requirements.txt
Django==4.2.7
psycopg2==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
tzdata==2023.3